backend/app/pyproject.toml
[project]
name = "johnkeats-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "google-genai",
    "google-cloud-firestore",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "websockets",
    "certifi",
]

[tool.setuptools]
packages = ["keats_agent", "tools"]
